Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float num;
-    scanf("%.4f",&num);
-    printf("You entered: %f\n",num);
+    double num;
+    scanf("%lf",&num);
+    printf("You entered: %lf\n",num);
     return 0;
 }
